Fix hexagonal cell rotation in `GridMap`
`GridMap` for square tiles needed to support 24 different orientations.
When hexagonal tile support was added, the hex tiles were being rotated
as they were cubes, 90 degrees on whichever axis.  This does not work
for hexagonal tiles.

Hex cell rotation is 60 degrees around the y-axis for 6 possible
orientations, then 180 degrees around the x-axis or z-axis (flipping the
tile over).  Any other rotation around x/z would cause a hex tile to
clip any neighboring cells. This results in 12 possible orientations for
hex cells in a GridMap.

In this commit, I added the hex-specific orientations to the GridMap,
along with the 60/180 degree rotations in the GridMap editor plugin.

In addition, I added `Math_SQRT3` as it is used within the hex
orientation.

Implementation wise, this adds a branch in multiple places to check if
the HexGrid uses square or hex cells, then use the appropriate basis
array.  If performance is a concern here, the two basis arrays can be
concatenated, but at a cost.  The cell orientation `Cell.rot` has 5
bits, allowing up to 32 orientations; square cells use 24 of those, and
hex cells add 12 more.

One option is to take away one bit from `Cell.item` or `Cell.layer`,
which I'm reluctant to do.  Anyone using square cells is going to lose
some flexibility to add hexagon support, which just feels bad.  This
would break backwards compatibility for GridMap::Cell values, resulting
in lost levels or incorrect tiles being displayed.

The other option allows us to fit the (24 + 12 = 36) possible
orientations into 32 options because four of the orientations for hexes
overlap with the square orientations. We could shove them all into the
same array, then reorganize the array such that the square & hex
orientations are overlapping, but contiguous segments; basically stick
the shared orientations in the middle of the array. That said, this also
breaks backwards compatibility, in a probably more horrific way; tiles
would be "randomly" rotated, but in a predictable/mappable way.