GridMap: use axial coordinate system for hex cells
The offset-based indexing approach offered many options for mapping
points into map indexes, but the current code had problems correctly
mapping various regions of a cell consistently
(https://github.com/godotengine/godot/pull/85890#issuecomment-2103848806).
Looking at the layout code, there were a lot of options with various
mathematical transforms for each one; verifying where the issue was, and
if it affected the other layouts was challenging.

There is also a comment at the top of the PR (#85890), and in the
discussion for the previous PR (#59842) about reducing the complexity of
this PR by removing the mixture of cell layout and orientation options.
That seemed like the best approach.

In this commit, I remove the existing layout & orientation options and
replace them with a single coordinate system (axial coordinates
described in
https://www.redblobgames.com/grids/hexagons/#coordinates-axial), and a
single cell orientation (pointy, previously the vertical orientation).

In addition I added tests for `GridMap::local_to_map()` and
`GridMap::map_to_local()` to verify we can correctly identify the edges
of the cell.

With this change we loose the ability to have irregularly shaped hexagon
cells where not all sides are the same length.  This is a side effect of
the current implementation of the algorithm to map between axial and 3d
coordinates.  I don't think this will be a big problem as most hex-based
3d games are using regular hexagons.