GridMapEditor: fix floorgrid offset issue
Currently if you exit editing a GridMap on floor != 0, when you return
to the GridMap the floor grid will be redrawn offset from the cusor.  So
the cursor is at floor 0, but the grid is drawn for floor -1.

The cause of this is `_draw_floor_grid()` takes into account the current
floor level when creating the floor grid.  The problem is that the grid
should be created at floor = 0, as it is offset by the floor number when
displayed.

This commit I removed the floor number argument from
`_draw_floor_grid()`.  As a result the floor grid is generated with
coordinates for floor == 0, and is correctly offset now.