HexGridEditor: fix for x-plane editing hex cells
With the conversion to axial coordinates for hex cells, editing cells
along the x-plane caused cells to be drawn in the wrong place.  It
turned out this was because `GridMapEditor::do_input_action()` assumed
it could modify the floor value directly after calling
`GridMap::local_to_map()`.  This looks like it was a workaround for
floating point errors that occur when raycasting against the picking
plane.

This causes a problem with hex cells as the axial coordinate system
means more math is required to set the x value.

In this commit I remove the code that modifies the floor/plane value for
the cell, and instead move the raypicking plane into the middle of the
cell plane. This will eliminate the majority of raypicking errors that
result from being directly on the edge of the cell.