GridMapEditor: Hex cell Q/R/S Axis editing support
Editing along the Z-axis with hex cells results in unexpected behavior
for users as cells will be placed in a zig-zag pattern.  This is because
with pointy orientation, the cell centers are offset every other row as
you move along the Z-axis.

In this commit, I added support for editing cells along the Q-axis
(northwest to southeast), and S-axis (southwest to northeast).  The
terms come from the axial coordinate system used in GridMap, and
derrived from:
https://www.redblobgames.com/grids/hexagons/#coordinates-axial

The Z and C hotkeys were repurposed to rotate around the various non-Y
axis options when using hex-shaped cells, using X, Q, Z, then R axis
when rotating clockwise (hotkey C), and the reverse for
counter-clockwise (hotkey Z). For square-shaped cells, these hotkeys
still map to the X-axis and Z-axis.

This commit does not include required changes to the `GridMap` editor
menu to show/hide the hex-specific axis.

Note, also this commit in particular removes code for functionality that
has been broken for some time.  PR #25406 Added support for using
Shift-Q and Shift-E to expand selection regions to other planes/floors
without changing the actual floor.  This does not function in 4.2.2.  It
looks like at some point the keybindings for change level were changed
from an accelerator to a shortcut, and the code for checking for the key
combination was not updated.

As the functionality does not currently work, and it's non-trivial to
reimplement with hex cells, I'm leaving it out for the time being.
A dedicated PR should be opened to expand selection capabilities within
the GridMapEditor later to re-introduce the functonality.