GridMap: Add get_cell_neighbors()
With the transition to axial coordinates for hex shaped cells in
GridMap, it became more difficult to navigate grid map cells to their
neighbors.  This is because the Q/R axial coordinates do not directly
map to x/y equivalents.

In this commit I add GridMap::get_cell_neighbors() which is the minimum
implementation required to map from a GridMap to an AStar3d without
duplicating the axial coordinate math.

This commit should be expanded on to add direction-based neighbor
functions, but I'd rather do that work in a dedicated PR to abstract
navigating cell indexes into a dedicated class for each cell type.