GridMapEditor: selection on Q/S planes fixed
Selection along the Q & S planes was selecting a cube volume due to
`GridMap::local_region_to_map()` using the two points as the corners of
an AABB.  This isn't a problem when editing on X/Y/Z planes as the begin
and end points share one common value for at least one of their indices.
The Q & S planes on the other hand, run diagonal across the X/Z
coordinates, so every selection is a cube traversing X/Z.

In this commit I add support for selecting along the Q/S plane.  To do
this, GridMapEditor will now prune the selection region returned from
`GridMap::local_region_to_map()` based on which axis is being edited.

In addition, I trimmed out a lot of extranous values & logic that
weren't being used in the selection code.